requirements.txt
flask==2.3.2
flask_sqlalchemy==3.0.3
python-dotenv==1.0.0
openai==0.27.0
pytest==7.4.0
Werkzeug<3.0.0
flask-wtf==1.1.1
